Refactor package structure to organize Google, GitHub, OpenWeather, Spotify, and Steam properties under 'tools' subpackage
