+++ src/main/java/space/parzival/discord/badley/configuration/properties/tools/SpotifyProperties.java
@@ -1,4 +1,4 @@
-package space.parzival.discord.badley.configuration.properties;
+package space.parzival.discord.badley.configuration.properties.tools;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
